feat(products): Implement authorization for product management operations

+++ app/Services/ProductService.php
@@ -6,17 +6,21 @@ use App\DTOs\Products\CreateProductData;
 use App\DTOs\Products\UpdateProductData;
 use App\DTOs\Stocks\CreateStockData;
 use App\Entities\Models\Product;
+use App\Policies\V1\ProductPolicy;
 use App\Repositories\Contracts\ProductRepositoryInterface;
 use App\Repositories\Contracts\StockRepositoryInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Facades\DB;
 
-class ProductService
+class ProductService extends BaseService
 {
     protected $productRepository;
 
     protected $stockRepository;
+
+    protected $policyClass = ProductPolicy::class;
 
     public function __construct(
         ProductRepositoryInterface $productRepository,
@@ -28,6 +32,10 @@ class ProductService
 
     public function createProduct(CreateProductData $productData): Product
     {
+        if (! $this->isAble('store', Product::class)) {
+            throw new AuthorizationException('You do not have permission to create a product.');
+        }
+
         return DB::transaction(function () use ($productData) {
 
             $product = $this->productRepository->create($productData);
@@ -59,11 +67,19 @@ class ProductService
 
     public function updateProduct(UpdateProductData $productData): Product
     {
+        if (! $this->isAble('update', Product::class)) {
+            throw new AuthorizationException('You do not have permission to update a product.');
+        }
+
         return $this->productRepository->update($productData);
     }
 
     public function deleteProduct(Product $product): bool
     {
+        if (! $this->isAble('delete', Product::class)) {
+            throw new AuthorizationException('You do not have permission to delete a product.');
+        }
+
         return $this->productRepository->delete($product);
     }
 }
